+++ sql/scripts/drop-mysql.sql
@@ -1,7 +1,0 @@
-
--- This is an example table drop script for MySQL. It is not required for running the unit tests, which use H2.
-
-drop table vertexium_vertex;
-drop table vertexium_edge;
-drop table vertexium_metadata;
-drop table vertexium_streaming_properties;