+++ core/src/main/java/org/vertexium/ElementBuilder.java
@@ -41,7 +41,7 @@ public abstract class ElementBuilder<T extends Element> implements ElementMutati
      * @param visibility The visibility to give this property.
      */
     public ElementBuilder<T> setProperty(String name, Object value, Visibility visibility) {
-        return setProperty(name, value, new Metadata(), visibility);
+        return setProperty(name, value, new Metadata(FetchHints.ALL), visibility);
     }
 
     /**
@@ -74,7 +74,7 @@ public abstract class ElementBuilder<T extends Element> implements ElementMutati
      * @param visibility The visibility to give this property.
      */
     public ElementBuilder<T> addPropertyValue(String key, String name, Object value, Visibility visibility) {
-        return addPropertyValue(key, name, value, new Metadata(), visibility);
+        return addPropertyValue(key, name, value, new Metadata(FetchHints.ALL), visibility);
     }
 
     /**