+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v4.0.0
 * Changed: FetchHints to support more filtering
+* Changed: Throw errors when calling methods without the proper fetch hints
 * Changed: Accumulo: Store metadata in an indexed list to prevent duplication in memory and over the wire 
 * Fixed: Accumulo Iterator memory leak 
 * Added: Method to Vertex to get more detailed edge summary