+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed: Accumulo Iterator memory leak 
 * Added: Method to Vertex to get more detailed edge summary
 * Removed: SQL
+* Removed: Blueprints and Titan support
 
 # v3.2.3
 * Fixed: Not adding all of the geoshape fields to index for existing elements